+++ components/dashboard/src/workspaces/WorkspaceEntry.tsx
@@ -9,8 +9,9 @@ import { GitpodHostUrl } from '@gitpod/gitpod-protocol/lib/util/gitpod-host-url'
 import ContextMenu, { ContextMenuEntry } from '../components/ContextMenu';
 import moment from 'moment';
 import Modal from '../components/Modal';
-import { MouseEvent, useState } from 'react';
+import { useState } from 'react';
 import { WorkspaceModel } from './workspace-model';
+import PendingChangesDropdown from '../components/PendingChangesDropdown';
 import Tooltip from '../components/Tooltip';
 
 function getLabel(state: WorkspaceInstancePhase) {
@@ -26,16 +27,7 @@ interface Props {
 
 export function WorkspaceEntry({ desc, model, isAdmin, stopWorkspace }: Props) {
     const [isModalVisible, setModalVisible] = useState(false);
-    const [isChangesModalVisible, setChangesModalVisible] = useState(false);
     const state: WorkspaceInstancePhase = desc.latestInstance?.status?.phase || 'stopped';
-    const pendingChanges = getPendingChanges(desc.latestInstance);
-    const numberOfChanges = pendingChanges.reduceRight((i, c) => i + c.items.length, 0)
-    let changesLabel = 'No Changes';
-    if (numberOfChanges === 1) {
-        changesLabel = '1 Change';
-    } else if (numberOfChanges > 1) {
-        changesLabel = numberOfChanges + ' Changes';
-    }
     const currentBranch = desc.latestInstance?.status.repo?.branch || Workspace.getBranchName(desc.workspace) || '<unknown>';
     const ws = desc.workspace;
     const startUrl = new GitpodHostUrl(window.location.href).with({
@@ -88,10 +80,6 @@ export function WorkspaceEntry({ desc, model, isAdmin, stopWorkspace }: Props) {
         );
     }
     const project = getProject(ws);
-    const showChanges = (event: MouseEvent) => {
-        event.preventDefault();
-        setChangesModalVisible(true);
-    }
     return <div>
         <div className="rounded-xl whitespace-nowrap flex space-x-2 py-6 px-6 w-full justify-between hover:bg-gray-100 dark:hover:bg-gray-800 focus:bg-gitpod-kumquat-light group">
             <div className="pr-3 self-center">
@@ -109,16 +97,9 @@ export function WorkspaceEntry({ desc, model, isAdmin, stopWorkspace }: Props) {
                     </a>
                 </div>
             </div>
-            <div className="flex w-2/12 truncate" onClick={numberOfChanges > 0 ? showChanges : undefined}>
-                <div className="flex flex-col">
-                    <div className="text-gray-500 truncate">{currentBranch}</div>
-                    {
-                        numberOfChanges > 0 ?
-                            <div className={"text-sm text-red-600 truncate cursor-pointer bg-red-50 group-hover:bg-red-100 hover:text-red-800 px-1.5 py-0.5 relative rounded-md -top-0.5"} onClick={showChanges}>{changesLabel}</div>
-                            :
-                            <div className="text-sm text-gray-400 truncate ">No Changes</div>
-                    }
-                </div>
+            <div className="flex flex-col items-start w-2/12">
+                <div className="text-gray-500 truncate">{currentBranch}</div>
+                <PendingChangesDropdown workspaceInstance={desc.latestInstance} />
             </div>
             <div className="flex w-2/12 self-center">
                 <Tooltip content={`Created ${moment(desc.workspace.creationTime).fromNow()}`}>
@@ -131,9 +112,6 @@ export function WorkspaceEntry({ desc, model, isAdmin, stopWorkspace }: Props) {
                 </ContextMenu>
             </div>
         </div>
-        <Modal visible={isChangesModalVisible} onClose={() => setChangesModalVisible(false)}>
-            {getChangesPopup(pendingChanges)}
-        </Modal>
         {isModalVisible && <Modal visible={isModalVisible} onClose={() => setModalVisible(false)}>
             <div>
                 <h3 className="pb-2">Delete Workspace</h3>
@@ -155,53 +133,12 @@ export function WorkspaceEntry({ desc, model, isAdmin, stopWorkspace }: Props) {
     </div>;
 }
 
-export interface PendingChanges {
-    message: string, items: string[]
-}
-
-export function getPendingChanges(wsi?: WorkspaceInstance): PendingChanges[] {
-    const pendingChanges: { message: string, items: string[] }[] = [];
-    const repo = wsi?.status.repo;
-    if (repo) {
-        if (repo.totalUncommitedFiles || 0 > 0) {
-            pendingChanges.push({
-                message: repo.totalUncommitedFiles === 1 ? 'an uncommited file' : `${repo.totalUncommitedFiles} uncommited files`,
-                items: repo.uncommitedFiles || []
-            });
-        }
-        if (repo.totalUntrackedFiles || 0 > 0) {
-            pendingChanges.push({
-                message: repo.totalUntrackedFiles === 1 ? 'an untracked file' : `${repo.totalUntrackedFiles} untracked files`,
-                items: repo.untrackedFiles || []
-            });
-        }
-        if (repo.totalUnpushedCommits || 0 > 0) {
-            pendingChanges.push({
-                message: repo.totalUnpushedCommits === 1 ? 'an unpushed commit' : `${repo.totalUnpushedCommits} unpushed commits`,
-                items: repo.unpushedCommits || []
-            });
-        }
-    }
-    return pendingChanges;
-}
-
 export function getProject(ws: Workspace) {
     if (CommitContext.is(ws.context)) {
         return `${ws.context.repository.host}/${ws.context.repository.owner}/${ws.context.repository.name}`;
     } else {
         return undefined;
     }
-}
-
-export function getChangesPopup(changes: PendingChanges[]) {
-    return <div className="flex flex-col space-y-4 w-96">
-        {changes.map(c => {
-            return <div className="">
-                <div className="text-gray-500">{c.message}</div>
-                {c.items.map(i => <div className="text-gray-400 text-xs">{i}</div>)}
-            </div>;
-        })}
-    </div>
 }
 
 export function WorkspaceStatusIndicator({instance}: {instance?: WorkspaceInstance}) {
